+++ examples/app-crm-minimal/src/providers/data/fetch-wrapper.ts
@@ -1,0 +1,58 @@
+import nookies from "nookies";
+
+type Error = {
+    message: string;
+    statusCode: string;
+};
+
+const customFetch = async (url: string, options: RequestInit) => {
+    const cookies = nookies.get();
+    const accessToken = cookies?.["access_token"];
+    const headers = options.headers as Record<string, string>;
+
+    return await fetch(url, {
+        ...options,
+        headers: {
+            ...headers,
+            Authorization: headers?.Authorization || `Bearer ${accessToken}`,
+            "Content-Type": "application/json",
+            "Apollo-Require-Preflight": "true",
+        },
+    });
+};
+
+export const fetchWrapper = async (url: string, options: RequestInit) => {
+    const response = await customFetch(url, options);
+
+    const responseClone = response.clone();
+    const body = await responseClone.json();
+    const error = getGraphQLErrors(body);
+
+    if (error) {
+        throw error;
+    }
+
+    return response;
+};
+
+const getGraphQLErrors = (body: any): Error | null => {
+    if (!body) {
+        return {
+            message: "Unknown error",
+            statusCode: "INTERNAL_SERVER_ERROR",
+        };
+    }
+
+    if ("errors" in body) {
+        const errors = body?.errors;
+        const messages = errors?.map((error: any) => error?.message)?.join("");
+        const code = errors?.[0]?.extensions?.code;
+
+        return {
+            message: messages || JSON.stringify(errors),
+            statusCode: code || 500,
+        };
+    }
+
+    return null;
+};