+++ packages/codemod/src/transformations/refine2-to-refine3.ts
@@ -599,7 +599,6 @@ const updateSetEditIdToSetId = (j: JSCodeshift, root: Collection<any>) => {
 
 const packagesToUpdate = [
     "@pankod/refine-airtable",
-    "@pankod/refine-altogic",
     "@pankod/refine-graphql",
     "@pankod/refine-hasura",
     "@pankod/refine-nestjsx-crud",