+++ packages/live-previews/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @refinedev/live-previews
 
+## 0.3.5
+
+### Patch Changes
+
+-   [#5114](https://github.com/refinedev/refine/pull/5114) [`00a9252c5de`](https://github.com/refinedev/refine/commit/00a9252c5de86aad544b0ca7d087c532c6d561fa) Thanks [@alicanerdurmaz](https://github.com/alicanerdurmaz)! - feat: `next` updated to latest version.
+
+-   Updated dependencies [[`00a9252c5de`](https://github.com/refinedev/refine/commit/00a9252c5de86aad544b0ca7d087c532c6d561fa)]:
+    -   @refinedev/antd@5.36.12
+
+## 0.3.4
+
+### Patch Changes
+
+-   [#5114](https://github.com/refinedev/refine/pull/5114) [`00a9252c5de`](https://github.com/refinedev/refine/commit/00a9252c5de86aad544b0ca7d087c532c6d561fa) Thanks [@alicanerdurmaz](https://github.com/alicanerdurmaz)! - feat: `next` updated to latest version.
+
+-   Updated dependencies [[`00a9252c5de`](https://github.com/refinedev/refine/commit/00a9252c5de86aad544b0ca7d087c532c6d561fa)]:
+    -   @refinedev/antd@5.36.11
+
 ## 0.3.3
 
 ### Patch Changes